artworkkkk manag

+++ src/Controller/BackOffice/BackCategoryController.php
@@ -1,30 +1,29 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\BackOffice;
 
 use App\Entity\Category;
 use App\Form\CategoryType;
 use App\Repository\CategoryRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
-#[Route('/category')]
-class CategoryController extends AbstractController
+#[Route('/back/category')]
+class BackCategoryController extends AbstractController
 {
-    #[Route('/', name: 'app_category_index', methods: ['GET'])]
+    #[Route('/', name: 'app_back_category_index', methods: ['GET'])]
     public function index(CategoryRepository $categoryRepository): Response
     {
-        return $this->render('category/index.html.twig', [
+        return $this->render('back_category/index.html.twig', [
             'categories' => $categoryRepository->findAll(),
         ]);
     }
 
-    #[Route('/new', name: 'app_category_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'app_back_category_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $category = new Category();
@@ -41,27 +40,27 @@ class CategoryController extends AbstractController
                 $entityManager->flush();
 
                 $this->addFlash('success', 'Category created successfully.');
-                return $this->redirectToRoute('app_category_index', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('app_back_category_index', [], Response::HTTP_SEE_OTHER);
             } catch (\Exception $e) {
                 $this->addFlash('error', 'Error creating category: ' . $e->getMessage());
             }
         }
 
-        return $this->render('category/new.html.twig', [
+        return $this->render('back_category/new.html.twig', [
             'category' => $category,
             'form' => $form,
         ]);
     }
 
-    #[Route('/{id}', name: 'app_category_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'app_back_category_show', methods: ['GET'])]
     public function show(Category $category): Response
     {
-        return $this->render('category/show.html.twig', [
+        return $this->render('back_category/show.html.twig', [
             'category' => $category,
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_category_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'app_back_category_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Category $category, EntityManagerInterface $entityManager): Response
     {
         $form = $this->createForm(CategoryType::class, $category);
@@ -76,30 +75,35 @@ class CategoryController extends AbstractController
                 $entityManager->flush();
 
                 $this->addFlash('success', 'Category updated successfully.');
-                return $this->redirectToRoute('app_category_index', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('app_back_category_index', [], Response::HTTP_SEE_OTHER);
             } catch (\Exception $e) {
                 $this->addFlash('error', 'Error updating category: ' . $e->getMessage());
             }
         }
 
-        return $this->render('category/edit.html.twig', [
+        return $this->render('back_category/edit.html.twig', [
             'category' => $category,
             'form' => $form,
         ]);
     }
 
-    #[Route('/{id}', name: 'app_category_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'app_back_category_delete', methods: ['POST'])]
     public function delete(Request $request, Category $category, EntityManagerInterface $entityManager): Response
     {
         if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($category);
-            $entityManager->flush();
+            if (count($category->getArtworks()) > 0) {
+                $this->addFlash('error', 'Cannot delete category that has artworks. Remove the artworks first.');
+            } else {
+                $entityManager->remove($category);
+                $entityManager->flush();
+                $this->addFlash('success', 'Category deleted successfully.');
+            }
         }
 
-        return $this->redirectToRoute('app_category_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_back_category_index');
     }
 
-    #[Route('/{id}/info', name: 'app_category_info', methods: ['GET'])]
+    #[Route('/{id}/info', name: 'app_back_category_info', methods: ['GET'])]
     public function getCategoryInfo(Request $request, ?Category $category = null): JsonResponse
     {
         try {
